#13 Pass all Goldbeter 1991 C++ tests
