+++ chaste_codegen_sbml/SBMLRefModels/src/srn/models/Tan2014/Tan2014OdeSystemAndSrnModel.cpp
@@ -1,4 +1,4 @@
-#include "Tan2014SrnModel.hpp"
+#include "Tan2014OdeSystemAndSrnModel.hpp"
 #include "CellwiseOdeSystemInformation.hpp"
 
 /* SBML ODE System */
