+++ chaste_codegen_sbml/SBMLRefModels/src/srn/models/VanLeeuwen2007/VanLeeuwen2007OdeSystemAndSrnModel.cpp
@@ -1,4 +1,4 @@
-#include "VanLeeuwen2007SrnModel.hpp"
+#include "VanLeeuwen2007OdeSystemAndSrnModel.hpp"
 #include "CellwiseOdeSystemInformation.hpp"
 
 /* SBML ODE System */
