+++ chaste_codegen_sbml/SBMLRefModels/src/srn/models/VanLeeuwen2007NonDim/VanLeeuwen2007NonDimOdeSystemAndSrnModel.cpp
@@ -1,4 +1,4 @@
-#include "VanLeeuwen2007NonDimSrnModel.hpp"
+#include "VanLeeuwen2007NonDimOdeSystemAndSrnModel.hpp"
 #include "CellwiseOdeSystemInformation.hpp"
 
 /* SBML ODE System */
