+++ chaste_codegen_sbml/SBMLRefModels/test/ContinuousTestPack.txt
@@ -1,6 +1,4 @@
-TestGoldbeter1991SrnModel/TestGoldbeter1991Srn.hpp
-TestGoldbeter1991SrnModel/TestGoldbeterOdeBasedSrnModels.hpp
-TestGoldbeter1991SrnModel/TestGoldbeterOdeSystem.hpp
+TestGoldbeter1991SBML.hpp
 TestTan2014SrnModel/TestTanOdeSystem.hpp
 TestVanLeeuwen2007SrnModel/TestVanLeeuwenOdeSystem.hpp
 TestVanLeeuwen2007NonDimSrnModel/TestVanLeeuwenNonDimOdeSystem.hpp
